Basic backend support for Safe Exam Browser (https://safeexambrowser.org/):
via the API is possible to associate a Safe Exam Browser configuration and its validation key to
the proctored object. The proctored page will then take care of enforcing that, if the safe exam
browser is in use, the supplied request header matches with the configuration and the URL.
When such match fails, we now return the configuration file to the browser, so that users can
open it and restart their session properly configured.